+++ src/Resources/contao/languages/de/default.php
@@ -2,3 +2,4 @@
 
 $GLOBALS['TL_LANG']['CTE']['internetschach_formular'] = array('Internetschach Formular', 'Formular Internetschach einfügen');
 $GLOBALS['TL_LANG']['CTE']['internetschach_anmeldungen'] = array('Internetschach Anmeldungen', 'Anmeldungen Internetschach einfügen');
+$GLOBALS['TL_LANG']['CTE']['internetschach_tabelle'] = array('Internetschach Tabelle', 'Tabelle Internetschach einfügen');