+++ src/Resources/contao/languages/de/tl_content.php
@@ -3,8 +3,30 @@
 $GLOBALS['TL_LANG']['tl_content']['internetschach_legend'] = 'Turnierserie';
 $GLOBALS['TL_LANG']['tl_content']['internetschach'] = array('Turnierserie wählen', 'Turnierserie wählen');
 
-$GLOBALS['TL_LANG']['tl_content']['internetschachdetails_legend'] = 'Turniere/Gruppen';
+$GLOBALS['TL_LANG']['tl_content']['internetschach_anmeldungen_legend'] = 'Turniere/Gruppen';
 $GLOBALS['TL_LANG']['tl_content']['internetschach_turniere'] = array('Turniere', 'Auszugebende Turniere wählen');
 $GLOBALS['TL_LANG']['tl_content']['internetschach_gruppen'] = array('Gruppen', 'Auszugebende Gruppen wählen');
 $GLOBALS['TL_LANG']['tl_content']['internetschach_viewturniere'] = array('Turniere anzeigen', 'Turnierespalte im Frontend anzeigen');
 $GLOBALS['TL_LANG']['tl_content']['internetschach_viewgruppen'] = array('Gruppen anzeigen', 'Gruppenspalte im Frontend anzeigen');
+
+$GLOBALS['TL_LANG']['tl_content']['internetschach_tabelle_legend'] = 'Tabelle';
+$GLOBALS['TL_LANG']['tl_content']['internetschach_tabelle'] = array('Anzuzeigende Tabelle', 'Tabelle, die im Frontend angezeigt werden sollen.');
+$GLOBALS['TL_LANG']['tl_content']['internetschach_spalten'] = array('Anzuzeigende Spalten', 'Spalten, die im Frontend angezeigt werden sollen. Sie werden in der unten angezeigten Reihenfolge ausgegeben.');
+
+$GLOBALS['TL_LANG']['tl_content']['internetschach_spalten_reference'] = array
+(
+	'platz'      => 'Pl.',
+	'cb-name'    => 'Benutzer',
+	'cb-land'    => 'Land',
+	'cb-rating'  => 'CBR',
+	'punkte'     => 'Pkt.',
+	'wertung1'   => 'Wtg1',
+	'wertung2'   => 'Wtg2',
+	'runden'     => '(Ergebnisse)',
+	'name'       => 'Name',
+	'dwz'        => 'DWZ',
+	'verein'     => 'Verein',
+	'fide-titel' => 'Titel',
+	'fide-elo'   => 'Elo',
+	'email'      => 'E-Mail'
+);